# filters/openpolicyagent: Add a sample benchmark with to measure percentiles