- Add benchmark for latency distribution with decision logging
- Fix a missed benchmark test case for race conditions