@mefarazath 2025-03-18
* Sample benchmark run with percentiles
```
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/filters/openpolicyagent/opaauthorizerequest
cpu: Apple M1 Max
BenchmarkWithPercentiles/minimal-with-decision-logs-10         	  127410	      8796 ns/op	   45762 B/op	     815 allocs/op
BenchmarkWithPercentiles/minimal-with-decision-logs/P50-10     	  127410	     59125 ns/op
BenchmarkWithPercentiles/minimal-with-decision-logs/P95-10     	  127410	    216125 ns/op
BenchmarkWithPercentiles/minimal-with-decision-logs/P99-10     	  127410	    529500 ns/op
BenchmarkWithPercentiles/minimal-with-decision-logs/P99.9-10   	  127410	   1463334 ns/op
PASS
```

* Results with `benchstat`
```
~ > benchstat minimal-percentiles-dl-old                                                                               10ms 
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/filters/openpolicyagent/opaauthorizerequest
cpu: Apple M1 Max
                                                    │ minimal-percentiles-dl-old │
                                                    │           sec/op           │
WithPercentiles/minimal-with-decision-logs-10                        52.35µ ± 1%
WithPercentiles/minimal-with-decision-logs/P50-10                    112.0µ ± 2%
WithPercentiles/minimal-with-decision-logs/P95-10                    702.5µ ± 2%
WithPercentiles/minimal-with-decision-logs/P99-10                    15.47m ± 1%
WithPercentiles/minimal-with-decision-logs/P99.9-10                  16.14m ± 1%
geomean                                                              1.006m

                                              │ minimal-percentiles-dl-old │
                                              │            B/op            │
WithPercentiles/minimal-with-decision-logs-10                 68.50Ki ± 0%

                                              │ minimal-percentiles-dl-old │
                                              │         allocs/op          │
WithPercentiles/minimal-with-decision-logs-10                  1.084k ± 0%
```

* Benchmark used to compare the performance of the new buffer implementation introduced by opa team in [PR](https://github.com/open-policy-agent/opa/pull/7446)


```
~ > benchstat minimal-decision-logs-percentiles-old minimal-decision-logs-percentiles-buffer-impl
goos: darwin
goarch: arm64
pkg: github.com/zalando/skipper/filters/openpolicyagent/opaauthorizerequest
cpu: Apple M1 Max
                                                    │ minimal-decision-logs-percentiles-old │ minimal-decision-logs-percentiles-buffer-impl │
                                                    │                sec/op                 │        sec/op          vs base                │
WithPercentiles/minimal-with-decision-logs-10                                  52.354µ ± 1%            8.887µ ±  6%  -83.03% (p=0.000 n=10)
WithPercentiles/minimal-with-decision-logs/P50-10                              112.02µ ± 2%            59.31µ ±  0%  -47.05% (p=0.000 n=10)
WithPercentiles/minimal-with-decision-logs/P95-10                               702.5µ ± 2%            220.5µ ± 14%  -68.61% (p=0.000 n=10)
WithPercentiles/minimal-with-decision-logs/P99-10                             15469.8µ ± 1%            543.8µ ± 12%  -96.48% (p=0.000 n=10)
WithPercentiles/minimal-with-decision-logs/P99.9-10                            16.143m ± 1%            1.499m ± 17%  -90.72% (p=0.000 n=10)
geomean                                                                         1.006m                 156.8µ        -84.41%

                                              │ minimal-decision-logs-percentiles-old │ minimal-decision-logs-percentiles-buffer-impl │
                                              │                 B/op                  │         B/op           vs base                │
WithPercentiles/minimal-with-decision-logs-10                            68.50Ki ± 0%            44.69Ki ± 0%  -34.76% (p=0.000 n=10)

                                              │ minimal-decision-logs-percentiles-old │ minimal-decision-logs-percentiles-buffer-impl │
                                              │               allocs/op               │       allocs/op        vs base                │
WithPercentiles/minimal-with-decision-logs-10                             1084.0 ± 0%              815.0 ± 0%  -24.82% (p=0.000 n=10)

```